requirements.txt
discord.py
Daug==2023.12.9.1
python-dotenv
redis
pytest-flake8
mastodon-py
